+++ history.md
@@ -3,6 +3,7 @@
 - only converts headers keys which are Symbols
 - use CGI for cookie parsing instead of custom code
 - unescape user and password before using them (patch provided by Lars Gierth)
+- expand ~ in ~/.restclientrc (patch provided by Mike Fletcher)
 
 # 1.5.0